requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-appconf==1.0.3
django-ckeditor==5.7.1
django-classy-tags==1.0.0
django-compressor==2.4
django-disqus==0.5
django-hitcount==1.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-sass-processor==0.8
django-tagging==0.4.6
django-taggit==1.1.0
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
guider==3.9.5
gunicorn==19.9.0
idna==2.8
libsass==0.19.4
Pillow==6.2.0
psycopg2==2.8.4
pytz==2019.2
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
utils==1.0.0
whitenoise==4.1.4
